refactor : [013-comment] UNIDENTIFIED_MEMBER_ID 을 상수화
Constants 클래스를 만들고, 기존의 domain/type 패키지를 domain/constants 로 네이밍 수정
