+++ src/main/java/com/trip/diary/domain/constants/Constants.java
@@ -1,0 +1,5 @@
+package com.trip.diary.domain.constants;
+
+public class Constants {
+    public static final Long UNIDENTIFIED_MEMBER_ID = -1L;
+}
