+++ src/main/java/com/trip/diary/domain/constants/MemberType.java
@@ -1,4 +1,4 @@
-package com.trip.diary.domain.type;
+package com.trip.diary.domain.constants;
 
 public enum MemberType {
     USER, ADMIN, DELETED;
