+++ src/main/java/com/trip/diary/domain/constants/ParticipantType.java
@@ -1,4 +1,4 @@
-package com.trip.diary.domain.type;
+package com.trip.diary.domain.constants;
 
 public enum ParticipantType {
     PENDING, ACCEPTED;
