+++ src/main/java/com/trip/diary/domain/model/Member.java
@@ -1,6 +1,6 @@
 package com.trip.diary.domain.model;
 
-import com.trip.diary.domain.type.MemberType;
+import com.trip.diary.domain.constants.MemberType;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
