+++ src/main/java/com/trip/diary/security/SecurityConfig.java
@@ -1,6 +1,6 @@
 package com.trip.diary.security;
 
-import com.trip.diary.domain.type.MemberType;
+import com.trip.diary.domain.constants.MemberType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.security.servlet.PathRequest;
 import org.springframework.context.annotation.Bean;
