+++ src/main/java/com/trip/diary/service/LocationService.java
@@ -16,7 +16,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static com.trip.diary.domain.type.ParticipantType.ACCEPTED;
+import static com.trip.diary.domain.constants.ParticipantType.ACCEPTED;
 import static com.trip.diary.exception.ErrorCode.NOT_AUTHORITY_READ_TRIP;
 
 @Service
